requirements.txt
streamlit
pdfplumber
python-docx
sentence-transformers
faiss-cpu
llama-cpp-python
transformers
blobfile
tiktoken